Remove experimental metric events
These definitions are a vestige of initial directory service metric
collection experiments and will most likely not be used in their current
form. The release of the directory service to the public has sparked
some privacy concerns about these definitions as well, so we might as
well delete them now to clarify their status.